Use modal overlay for application exit.
